Fix some logging code failed to work and refine log format
As is in issue #18, this commit fixes some logging code failed to work.

That's because `Logger.<debug|info|warning|error>()` has a similar but
different parameter form from that of `print()`. The method of `Logger`
needs the positional arguments other than the first one to be contents
of the `%`-like templates in the first `str` argument, but that's very
different from `print()`'s behavior.

And this commit refines format of some logging code.